editing a few Do's statements
I removed a few items which are important but I think are more appropriate for a section where we describe what we value as a group instead of the code of conduct where we are more focused on appropriate behavior. The ones I removed:

Take initiative in your learning and exploration
Be curious.

Any good place to add them on the website?